move duplicated createOutputFailedSet to base
