+++ imod/protocols/protocol_tomoReconstruction.py
@@ -26,11 +26,10 @@
 
 import os
 
-from pwem.emlib.image import ImageHandler
 from pyworkflow import BETA
 from pyworkflow.object import Set
 import pyworkflow.protocol.params as params
-from tomo.objects import Tomogram, TiltSeries, TiltImage
+from tomo.objects import Tomogram
 
 from .. import Plugin
 from .protocol_base import ProtImodBase, EXT_MRC_ODD_NAME, EXT_MRC_EVEN_NAME
@@ -312,36 +311,6 @@ class ProtImodTomoReconstruction(ProtImodBase):
 
             output.append(newTomogram)
             output.update(newTomogram)
-            output.write()
-            self._store()
-
-    def createOutputFailedSet(self, tsObjId):
-        # Check if the tilt-series ID is in the failed tilt-series
-        # list to add it to the set
-        if tsObjId in self._failedTs:
-            ts = self.inputSetOfTiltSeries.get()[tsObjId]
-            tsSet = self.inputSetOfTiltSeries.get()
-            tsId = ts.getTsId()
-
-            output = self.getOutputFailedSetOfTiltSeries(tsSet)
-
-            newTs = TiltSeries(tsId=tsId)
-            newTs.copyInfo(ts)
-            output.append(newTs)
-
-            for index, tiltImage in enumerate(ts):
-                newTi = TiltImage()
-                newTi.copyInfo(tiltImage, copyId=True, copyTM=True)
-                newTi.setAcquisition(tiltImage.getAcquisition())
-                newTi.setLocation(tiltImage.getLocation())
-                newTs.append(newTi)
-
-            ih = ImageHandler()
-            x, y, z, _ = ih.getDimensions(newTs.getFirstItem().getFileName())
-            newTs.setDim((x, y, z))
-            newTs.write(properties=False)
-
-            output.update(newTs)
             output.write()
             self._store()
